requirements.txt
numpy==1.26.4
paddlepaddle==2.5.2
paddleaudio==1.0.2